requirements.txt
aiogram==3.15.0
aiosqlite==0.20.0
apscheduler==3.10.4
aiohttp==3.10.11
beautifulsoup4==4.12.3
requests==2.32.3
pillow==10.4.0